[CCUBE-256][KT] Rename file
